[REFACTOR] QnA 수정 사항에 따른 수정

+++ src/main/java/com/example/kbuddy_backend/blog/dto/response/BlogPaginationResponse.java
@@ -2,10 +2,10 @@ package com.example.kbuddy_backend.blog.dto.response;
 
 import java.time.LocalDateTime;
 
-public record BlogPaginationResponse(Long id, Long writerId, Long categoryId, String title, String description,
+public record BlogPaginationResponse(Long id, Long writerId, int categoryId, String title, String description,
                                     int viewCount, int heartCount,
                                     int commentCount, LocalDateTime createdAt, LocalDateTime modifiedAt) {
-    public static BlogPaginationResponse of(Long id, Long writerId, Long categoryId, String title, String description,
+    public static BlogPaginationResponse of(Long id, Long writerId, int categoryId, String title, String description,
                                             int viewCount, int heartCount,
                                             int commentCount, LocalDateTime createdAt, LocalDateTime modifiedAt) {
         return new BlogPaginationResponse(id, writerId, categoryId, title, description, viewCount, heartCount,
